Refactor services
- refactor the service layer to be more general
- move forms closer to where they will be used
- changed names to be more concise
- register and login pages call backend
